Load config file and custom prompt.
This will be the last commit and most important commit for the weekend.
It adds support for a local config file, which is instance eval'd into
the Console object before the REPL is started, giving you the chance to
set anything you please.

Down the road I also plan to add support for global and local config
files but for now this will do nicely.

I also chose a simple custom prompt for Cage which shows the current
domain as well as the status code of the last response. This is of
course fully customizable in the config file.